SimpleMacro_Testing/requirements.txt
opencv-python==4.8.1.78
numpy==1.26.2
pyautogui==0.9.54
Pillow==10.1.0
keyboard==0.13.5
mss==9.0.1
pygetwindow==0.0.9
pynput==1.7.6
